TransistorElm.java: fix transistor accuracy problems
The class TransistorElm, which simulates bipolar junction transistors
(BJTs), had some model inaccuracies. It was supposed to follow the
Ebers-Moll model, the basic model of BJT behavior, but the gain
factors used to calculate currents were handled improperly: the
Ebers-Moll "reciprocity relation", which relates the saturation
currents of the P-N junctions to the forward and reverse current
gains, was not accounted for. This resulted in incorrect behavior
under some peculiar circumstances: a saturated high-gain transistor
with a very low collector-emitter voltage would direct most of its
base current out the emitter, rather than out the collector as it
should. Also the currents in this situation, as the collector-emitter
voltage approached zero, would be only half as large as the Ebers-Moll
model predicts.

In addition, for calculating currents, negative base-emitter voltages
had been effectively clamped to zero, in order to prevent a singular
matrix which could otherwise result from underflow. This clamping has
now been removed; singular matrices are prevented in a different way.
Starting from the very first subiteration of each timestep, a tiny
conductance "gmin" (only 1% of the "leakage" parameter) is added
parallel to each of the BJT's P-N junctions. Though this conductance
is too small to visibly affect the component's operation, it is large
enough to prevent the underflows that can create a singular matrix.
Since "gmin" is already used to help force convergence if the circuit
requires too many subiterations to converge, the only change here is
to begin with a small but nonzero value of "gmin".

The final change is to adjust the thermal voltage "vt" upward a bit,
to correspond to SPICE's default circuit temperature of 27 degrees
Celsius (300.15 Kelvin). The value previously used corresponds to 17
degrees Celsius, decidedly on the cool side. Since the terms in BJT
currents vary exponentially with "vt", even this small temperature
change can causes a substantial change in current for any given
voltage. A default temperature of 27 C or 300 K is standard for many
analog circuit simulations, and using a value consistent with it will
make it a little easier to check results against such simulators.