Make transformers reset properly.
This change fixes the reset behavior for both TransformerElm and
TappedTransformerElm. The reset() method for both of these classes
was not zeroing their current-source variables; this has now been
fixed.

(The following explanation is probably already familiar, since I
simply patterned this fix off of existing fixes already done for
capacitors and inductors. I'm including it for the benefit of
future readers of this commit.)

Suppose one of the transformer's nodes is the ground node. For
example, this can easily happen in a simple circuit with a
two-terminal voltage source directly connected to the transformer,
since if the circuit has no ground element, one of the voltage source
terminals will be assigned to the ground node. During the circuit's
initial analysis, setNodeVoltage() will be called for each circuit
element connected to the ground node, to ensure that node is set
to 0 volts. But setNodeVoltage() will in turn call the element's
calculateCurrent() method. For transformers, this method will set
the current variables using the existing values for the current-source
variables, just as it does for each integration step. If reset() does
not zero the current-source variables along with the current variables,
the current may thus end up with a non-zero value after reset. So to
reset the transformer properly, reset() must also zero its
current-source variables.

In addition to this fix, I did a small cleanup in TappedTransformerElm,
moving the setting of current[3] (which contains the *net* current
through the secondary's center tap) from inside the draw() method to
the end of the calculateCurrent() method, which seemed a more
appropriate place. Just to be safe and consistent, current[3] is also
zeroed on reset(), like the other current variables.