Fix clearing on model dialog
This commit fixes a bug in the composite model (ie subcircuit) editor dialog where the chip display is not cleared properly, resulting in buffer garbage.
This bug occurs when the model is large and the camera is automatically zoomed out, and is caused by the rendering context having an undefined transform when clearing the background. Resetting the transform before clearing fixes this bug.