WebSocket: Fire event on finished reloading of iframe
When reloading the page, listening just for the iframe "load" event is
insufficient, as there is a small window of time during which the iframe
load event has occured, but the simulator is not yet initialized. We
want to be sure that the client can execute anything without having to
double check, therefore we send an event only when we're sure the
simulator is back up and running.