WebSocket: implement automatic shutdown after WS close
The default behavior of the WebSocket shim is to continuously retry to
connect to the WS server. If that server is short-lived and goes away
(possibly during development), the simulator continues running in the
browser window. This option can be specified to instruct the WebSocket
code to never reconnect after the server has gone away and instead to
shut down the whole simulation by redirecting the iframe to about:blank.